refactor(api): move route handlers to lib modules to preserve unit testability
